Enable safeSearch option of continuation search
* Add UTIL#buildPostBody
* Remove unnecessary stuff
