+++ lib/main.js
@@ -8,18 +8,6 @@ const BASE_URL = 'https://www.youtube.com/';
 const BASE_SEARCH_URL = 'https://www.youtube.com/results?';
 const BASE_API_URL = 'https://www.youtube.com/youtubei/v1/search?key=';
 
-const DEFAULT_CONTEXT = {
-  client: {
-    utcOffsetMinutes: 0,
-    gl: 'US',
-    hl: 'en',
-    clientName: 'WEB',
-    clientVersion: '<important information>',
-  },
-  user: {},
-  request: {},
-};
-
 const main = module.exports = async(searchString, options) => {
   const opts = UTIL.checkArgs(searchString, options);
 
@@ -68,17 +56,15 @@ const main = module.exports = async(searchString, options) => {
   if (!token || opts.limit < 1) return resp;
 
   // Recursively fetch more items
-  const nestedResp = await parsePage2(parsed.apiKey, parsed.clientVersion, token, opts);
+  const nestedResp = await parsePage2(parsed.apiKey, token, opts);
 
   // Merge the responses
   resp.items.push(...nestedResp);
   return resp;
 };
 
-const parsePage2 = async(apiKey, clientVersion, token, opts) => {
-  DEFAULT_CONTEXT.client.clientVersion = clientVersion;
-
-  const json = await UTIL.doPost(BASE_API_URL + apiKey, opts, { continuation: token, context: DEFAULT_CONTEXT });
+const parsePage2 = async(apiKey, token, opts) => {
+  const json = await UTIL.doPost(BASE_API_URL + apiKey, opts, UTIL.buildPostBody(token, opts));
 
   const wrapper = json.onResponseReceivedCommands[0].appendContinuationItemsAction.continuationItems;
 
@@ -97,7 +83,7 @@ const parsePage2 = async(apiKey, clientVersion, token, opts) => {
   if (!nextToken || opts.limit < 1) return parsedItems;
 
   // Recursively fetch more items
-  const nestedResp = await parsePage2(apiKey, clientVersion, nextToken, opts);
+  const nestedResp = await parsePage2(apiKey, nextToken, opts);
   parsedItems.push(...nestedResp);
   return parsedItems;
 };
