+++ lib/util.js
@@ -1,9 +1,19 @@
-const OLD_UTIL = require('./util.js');
 const URL = require('url');
 
 const BASE_URL = 'https://www.youtube.com/';
 const DEFAULT_OPTIONS = { limit: 100, safeSearch: false, nextpageRef: null };
 const DEFAULT_QUERY = { gl: 'US', hl: 'en' };
+const DEFAULT_CONTEXT = {
+  client: {
+    utcOffsetMinutes: 0,
+    gl: 'US',
+    hl: 'en',
+    clientName: 'WEB',
+    clientVersion: '<important information>',
+  },
+  user: {},
+  request: {},
+};
 
 exports.parseFilters = json => {
   const wrapper = json.contents.twoColumnSearchResultsRenderer.primaryContents.sectionListRenderer;
@@ -34,8 +44,9 @@ exports.parseFilters = json => {
 
 exports.parseBody = body => {
   const str = between(body, 'var ytInitialData =', '; \n');
-  const apiKey = between(body, 'INNERTUBE_API_KEY":"', '"') || OLD_UTIL.between(body, 'innertubeApiKey":"', '"');
-  const clientVersion = between(body, 'INNERTUBE_CONTEXT_CLIENT_VERSION":"', '"') || OLD_UTIL.between(body, 'innertube_context_client_version":"', '"');
+  const apiKey = between(body, 'INNERTUBE_API_KEY":"', '"') || between(body, 'innertubeApiKey":"', '"');
+  const clientVersion = between(body, 'INNERTUBE_CONTEXT_CLIENT_VERSION":"', '"') || between(body, 'innertube_context_client_version":"', '"');
+  DEFAULT_CONTEXT.client.clientVersion = clientVersion;
   return { json: JSON.parse(str || null), apiKey, clientVersion };
 };
 
@@ -102,4 +113,14 @@ const between = exports.between = (haystack, left, right) => {
   if (pos === -1) { return ''; }
   haystack = haystack.slice(0, pos);
   return haystack;
+};
+
+exports.buildPostBody = (token, options) => {
+  let body = { context: DEFAULT_CONTEXT, continuation: token };
+  if (!options) return body;
+  if (options.gl) body.context.client.gl = options.gl;
+  if (options.hl) body.context.client.hl = options.hl;
+  if (options.utcOffsetMinutes) body.context.client.utcOffsetMinutes = options.utcOffsetMinutes;
+  if (options.safeSearch) body.context.user.enableSafetyMode = true;
+  return body;
 };
